% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pw.R
\name{calculate_pw_and_dpw}
\alias{calculate_pw_and_dpw}
\title{Calculate derivative of pw}
\usage{
calculate_pw_and_dpw(SG, y, logtheta, return_lS = FALSE)
}
\arguments{
\item{SG}{SGGP object}

\item{y}{Measured values for SG$design}

\item{logtheta}{Log of correlation parameters}

\item{return_lS}{Should lS be returned?}
}
\value{
derivative matrix of pw with respect to logtheta
}
\description{
Calculate derivative of pw
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
calculate_pw_and_dpw(SG=SG, y=y, logtheta=c(-.1,.1,.3))
}
