% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fastcalcassist_fs.R
\name{calculate_pw_and_dpw_C}
\alias{calculate_pw_and_dpw_C}
\title{Calculate derivative of pw}
\usage{
calculate_pw_and_dpw_C(SG, y, theta, return_lS = FALSE)
}
\arguments{
\item{return_dlS}{Should dlS be returned?}
}
\value{
derivative matrix of pw with respect to logtheta
}
\description{
Calculate derivative of pw
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
calculate_pw_and_dpw_C(SG=SG, y=y, logtheta=c(-.1,.1,.3))
}
