% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pw_C.R
\name{calculate_pw_C}
\alias{calculate_pw_C}
\title{Calculate predictive weights for SGGP}
\usage{
calculate_pw_C(SG, y, logtheta, return_lS = FALSE)
}
\arguments{
\item{SG}{SGGP object}

\item{y}{Measured values for SG$design}

\item{logtheta}{Log of correlation parameters}

\item{return_lS}{Should lS be returned?}
}
\value{
Vector with predictive weights
}
\description{
Predictive weights are Sigma^{-1}*y in standard GP.
This calculation is much faster since we don't need to
solve the full system of equations.
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
calculate_pw(SG=SG, y=y, logtheta=c(-.1,.1,.3))
}
