% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGGP.R
\docType{package}
\name{CGGP}
\alias{CGGP}
\title{CGGP: A package for running sparse grid computer experiments}
\description{
The CGGP package implements the method presented in Plumlee et al. (2019).
}
\section{CGGP functions}{

The CGGP functions: CGGPcreate, CGGPfit, CGGPappend, and CGGPpred
}

