% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGGP_append_fs.R
\name{CGGP_internal_calcMSE}
\alias{CGGP_internal_calcMSE}
\title{Calculate MSE over single dimension}
\usage{
CGGP_internal_calcMSE(xl, theta, CorrMat)
}
\arguments{
\item{xl}{Vector of points in 1D}

\item{theta}{Correlation parameters}

\item{CorrMat}{Function that gives correlation matrix for vectors of 1D points.}
}
\value{
MSE value
}
\description{
Calculated using grid of integration points.
Can be calculated exactly, but not much reason in 1D.
}
\examples{
CGGP_internal_calcMSE(xl=c(0,.5,.9), theta=c(1,2,3),
         CorrMat=CGGP_internal_CorrMatCauchySQT)
}
