% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGGP_pred_fs.R
\name{CGGP_internal_MSEpredcalc}
\alias{CGGP_internal_MSEpredcalc}
\title{Calculate MSE prediction along a single dimension}
\usage{
CGGP_internal_MSEpredcalc(xp, xl, theta, CorrMat)
}
\arguments{
\item{xp}{Points at which to calculate MSE}

\item{xl}{Levels along dimension, vector???}

\item{theta}{Correlation parameters}

\item{CorrMat}{Function that gives correlation matrix for vectors of 1D points.}
}
\value{
MSE predictions
}
\description{
Calculate MSE prediction along a single dimension
}
\examples{
CGGP_internal_MSEpredcalc(c(.4,.52), c(0,.25,.5,.75,1), theta=c(.1,.2),
             CorrMat=CGGP_internal_CorrMatCauchySQ)
}
