% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGGP_speedup_postvarcalc.R
\name{CGGP_internal_postvarmatcalcfaster}
\alias{CGGP_internal_postvarmatcalcfaster}
\title{Calculate posterior variance, faster version}
\usage{
CGGP_internal_postvarmatcalcfaster(GMat, dGMat, cholS, dSMat, INDSN,
  numpara, ..., returnlogs = FALSE, returnderiratio = FALSE,
  returndG = FALSE, returndiag = FALSE)
}
\arguments{
\item{GMat}{Matrix}

\item{dGMat}{Derivative of matrix}

\item{cholS}{Cholesky factorization of S}

\item{dSMat}{Deriv of SMat}

\item{INDSN}{Indices, maybe}

\item{numpara}{Number of parameters for correlation function}

\item{...}{Placeholder}

\item{returnlogs}{Should log scale be returned}

\item{returnderiratio}{Should derivative ratio be returned?}

\item{returndG}{Should dG be returned}

\item{returndiag}{Should diag be returned}
}
\value{
Variance posterior
}
\description{
Calculate posterior variance, faster version
}
\examples{
CGGP_internal_postvarmatcalc(c(.4,.52), c(0,.25,.5,.75,1),
             xo=c(.11), theta=c(.1,.2,.3),
             CorrMat=CGGP_internal_CorrMatCauchySQT)
}
