% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGGP_sharedfunctions.R
\name{CGGP_internal_set_corr}
\alias{CGGP_internal_set_corr}
\title{Set correlation function of CGGP object}
\usage{
CGGP_internal_set_corr(CGGP, corr)
}
\arguments{
\item{CGGP}{CGGP object}

\item{corr}{Correlation function}
}
\value{
CGGP object
}
\description{
Set correlation function of CGGP object
}
\examples{
obj <- CGGPcreate(3, 20, corr="matern52")
CGGP_internal_set_corr(obj, "gaussian")
}
