% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGGP_append_fs.R
\name{CGGPappend}
\alias{CGGPappend}
\title{Add points to CGGP}
\usage{
CGGPappend(CGGP, batchsize, selectionmethod = "UCB",
  multioutputdim_weights = 1)
}
\arguments{
\item{CGGP}{Sparse grid object}

\item{batchsize}{Number of points to add}

\item{selectionmethod}{How points will be selected: one of `UCB`, `TS`,
`Greedy`, `Oldest`, `Random`, or `Lowest`}

\item{multioutputdim_weights}{Weights for each output dimension.}
}
\value{
SG with new points added.
}
\description{
Add `batchsize` points to `SG` using `theta`.
}
\examples{
SG <- CGGPcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2})
SG <- CGGPfit(SG, Y=y)
SG <- CGGPappend(CGGP=SG, batchsize=20)
# UCB,TS,Greedy
}
\seealso{
Other CGGP core functions: \code{\link{CGGPcreate}},
  \code{\link{CGGPfit}}, \code{\link{predict.CGGP}}
}
\concept{CGGP core functions}
