% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPstuff.R
\name{CorrMat}
\alias{CorrMat}
\title{Calculate correlation matrix for two sets of points in one dimension}
\usage{
CorrMat(x1, x2, .., logtheta, theta)
}
\arguments{
\item{x1}{Vector of coordinates from same dimension}

\item{x2}{Vector of coordinates from same dimension}

\item{theta}{Is this log theta???}
}
\value{
Matrix
}
\description{
Note that this isn't the correlation between two vectors.
It is for two
}
\examples{
CorrMat(c(0,.2,.4),c(.1,.3,.5), 10)
}
