% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPstuff.R
\name{CorrMatMatern32}
\alias{CorrMatMatern32}
\title{Calculate correlation matrix for two sets of points in one dimension}
\usage{
CorrMatMatern32(x1, x2, ..., logtheta, theta)
}
\arguments{
\item{x1}{Vector of coordinates from same dimension}

\item{x2}{Vector of coordinates from same dimension}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta}{Log of correlation parameter}

\item{theta}{Correlation parameter}
}
\value{
Matrix
}
\description{
Note that this isn't the correlation between two vectors.
It is for two
}
\examples{
CorrMatMatern32(c(0,.2,.4),c(.1,.3,.5), theta=.1)
}
