% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPstuff.R
\name{dCorrMat}
\alias{dCorrMat}
\title{Calculate derivative of CorrMat with respect to theta}
\usage{
dCorrMat(x1, x2, theta)
}
\arguments{
\item{x1}{First vector of 1D points}

\item{x2}{Second vector of 1D points}

\item{theta}{Actually log theta, terrible notation}
}
\value{
Matrix
}
\description{
Calculate derivative of CorrMat with respect to theta
}
\examples{
dCorrMat(c(0,.2,.4),c(.1,.3,.5), 10)
}
