% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrMatern32.R
\name{ddiag_corrMatMatern32}
\alias{ddiag_corrMatMatern32}
\title{Derivative of diagonal of Matern 3/2 correlation}
\usage{
ddiag_corrMatMatern32(x1, ..., logtheta, theta, nugget)
}
\arguments{
\item{x1}{Vector of 1D points}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta}{Log of correlation parameters}

\item{theta}{Correlation parameters}

\item{nugget}{Nugget to add to each correlation.}
}
\value{
Vector
}
\description{
It's just 0.
}
\examples{
ddiag_corrMatMatern32(c(0,.2,.4,.6,.8,1),theta=.1, nugget=0) # Should be all zeros
}
