% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPstuff.R
\name{diag_corrMatMatern32}
\alias{diag_corrMatMatern32}
\title{Calculate correlation of points with themselves}
\usage{
diag_corrMatMatern32(x1, ..., logtheta, theta, nugget)
}
\arguments{
\item{x1}{Vector of 1D points}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta}{Log of correlation parameters}

\item{theta}{Correlation parameters}

\item{nugget}{Nugget to add to each correlation.}
}
\value{
Correlations of x1 with itself
}
\description{
Use to get diagonal of correlation matrix.
}
\examples{
diag_corrMatMatern32(c(0,.2,.4,.6,.8,1),theta=.1, nugget=0) # Should be all ones
}
