% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrGauss.R
\name{diag_gausscorr}
\alias{diag_gausscorr}
\title{Gaussian correlation matrix}
\usage{
diag_gausscorr(x1, theta, logtheta, nugget)
}
\arguments{
\item{x1}{Vector of 1D points}

\item{theta}{Correlation parameter in 1D, one number}

\item{logtheta}{Log of correlation parameters}

\item{nugget}{Nugget to add to each correlation.}
}
\value{
Correlation matrix
}
\description{
Gaussian correlation matrix
}
\examples{
gausscorr(c(0,.5,1), c(.1,.4,.5,.9), theta=.1, nugget=0)
}
