% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPpredstuff.R
\name{dMSEpred_calc}
\alias{dMSEpred_calc}
\title{Derivative of MSE calculation}
\usage{
dMSEpred_calc(xp, xl, ..., logtheta, theta, nugget, CorrMat, diag_corrMat,
  dCorrMat, ddiag_corrMat)
}
\arguments{
\item{xp}{Points at which to calculate MSE}

\item{xl}{Levels along dimension}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta}{Log of correlation parameters}

\item{theta}{Correlation parameters}

\item{nugget}{Nugget to add to diagonal of correlation matrix}

\item{CorrMat}{Function that gives correlation matrix for vectors of 1D points.}

\item{diag_corrMat}{Function that gives diagonal of correlation matrix}

\item{dCorrMat}{Derivative of CorrMat}

\item{ddiag_corrMat}{Derivative of diagonal of diag_corrMat
for vector of 1D points.}
}
\value{
MSE predictions
}
\description{
Derivative of MSE calculation
}
\examples{
MSEpred_calc(c(.4,.52), c(0,.25,.5,.75,1), theta=.1, nugget=1e-5,
             CorrMat=CorrMatMatern32,
             diag_corrMat=diag_corrMatMatern32)
}
