% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrGauss.R
\name{gausscorr}
\alias{gausscorr}
\title{Gaussian correlation matrix}
\usage{
gausscorr(x1, x2, theta)
}
\arguments{
\item{x1}{Vector of 1D points}

\item{x2}{Vector of 1D points}

\item{theta}{Correlation parameter in 1D, one number}
}
\value{
Correlation matrix
}
\description{
Gaussian correlation matrix
}
\examples{
gausscorr(c(0,.5,1), c(.1,.4,.5,.9), theta=.1)
}
