% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPlik.R
\name{glik}
\alias{glik}
\title{Gradient of likelihood. Is it log likelihood?}
\usage{
glik(logtheta, ..., SG, y, return_lik = FALSE)
}
\arguments{
\item{logtheta}{Log of correlation parameters}

\item{...}{Don't use, just forces theta to be named}

\item{SG}{SGGP object}

\item{y}{SG$design measured values}

\item{return_lik}{If yes, it returns a list with lik and glik}
}
\value{
Vector for gradient of likelihood w.r.t. x (theta)
}
\description{
Gradient of likelihood. Is it log likelihood?
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
glik(c(.1,.1,.1), SG=SG, y=y)
}
