% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPlik.R
\name{glikMV}
\alias{glikMV}
\title{Gradient of likelihood for multivariate SGGP}
\usage{
glikMV(logtheta, SG, yMV)
}
\arguments{
\item{logtheta}{log of correlation parameters}

\item{SG}{SGGP object}

\item{yMV}{Matrix with output, number of columns is number of outputs}
}
\value{
Vector, Gradient of likelihood of logtheta
}
\description{
Gradient of likelihood for multivariate SGGP
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y1 <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
y2 <- apply(SG$design, 1, function(x){x[1]^1.3+.4*sin(6*x[2])+rnorm(1,0,.01)})
y <- cbind(y1, y2)
glikMV(logtheta=c(.1,.2,.3), SG=SG, yMV=y)
}
