% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPvalid.R
\name{gvalidation}
\alias{gvalidation}
\title{Calculate gradient of validation with respect to logtheta}
\usage{
gvalidation(logtheta, SG, y, xval, yval)
}
\arguments{
\item{logtheta}{Log of correlation parameters}

\item{SG}{SGGP object}

\item{y}{Measured values of SG$design}

\item{xval}{Validation X matrix}

\item{yval}{Validation y vector}
}
\value{
Vector, gradient of `validation`
}
\description{
Calculate gradient of validation with respect to logtheta
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
f1 <- function(x){x[1]+x[2]^2+rnorm(1,0,.01)}
y <- apply(SG$design, 1, f1)
Xval <- matrix(runif(3*100), ncol=3)
Yval <- apply(Xval, 1, f1)
gvalidation(c(.1,.1,.1), SG=SG, y=y, xval=Xval, yval=Yval)
}
