% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPlik.R
\name{lik}
\alias{lik}
\title{Calculate likelihood}
\usage{
lik(logtheta, ..., SG, y)
}
\arguments{
\item{logtheta}{Log of correlation parameters}

\item{...}{Don't use, just forces theta to be named}

\item{SG}{SGGP object}

\item{y}{Measured values of SG$design}
}
\value{
Likelihood
}
\description{
Calculate likelihood
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
lik(c(.1,.1,.1), SG=SG, y=y)
}
