% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPappendstuff.R
\name{MSE_de}
\alias{MSE_de}
\title{Calculate MSE over blocks}
\usage{
MSE_de(valsinds, MSE_v)
}
\arguments{
\item{valsinds}{Block levels to calculate MSEs for}

\item{MSE_v}{Matrix of MSE values}
}
\value{
All MSE values
}
\description{
Delta of adding block is product over i=1..d of IMSE(i,j-1) - IMSE(i,j)
}
\examples{
SG <- SGcreate(c(0,0,0), c(1,1,1), batchsize=100)
theta <- c(.1,.1,.1)
MSE_v <- outer(1:SG$d, 1:8, 
     Vectorize(function(lcv1, lcv2) {
        MSE_calc(SG$xb[1:SG$sizest[lcv2]], theta=theta[lcv1])
 }))
MSE_de(SG$po[1:SG$poCOUNT, ], MSE_v)
}
