% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrGauss.R
\name{MSEgauss}
\alias{MSEgauss}
\title{Calculate MSE for Gaussian correlation}
\usage{
MSEgauss(xl, theta)
}
\arguments{
\item{xl}{Vector of evaluated points in 1D}

\item{theta}{Correlation parameter, one number}
}
\value{
MSE
}
\description{
Calculate MSE for Gaussian correlation
}
\examples{
th = .1
xl <- c(0,.5,.9)
MSE_calc.out <- MSEgauss(xl=xl, theta=th)
}
