% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{neglogpost}
\alias{neglogpost}
\title{Calculate negative posterior}
\usage{
neglogpost(theta, SG, y)
}
\arguments{
\item{theta}{Correlation parameters}

\item{SG}{SGGP object}

\item{y}{Measured values of SG$design}

\item{...}{Don't use, just forces theta to be named}
}
\value{
Likelihood
}
\description{
Calculate negative posterior
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
lik(c(.1,.1,.1), SG=SG, y=y)
}
