% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGGP_S3_methods.R
\name{plot.CGGP}
\alias{plot.CGGP}
\title{S3 plot method for CGGP}
\usage{
\method{plot}{CGGP}(x, y, ...)
}
\arguments{
\item{x}{CGGP object}

\item{y}{Don't use}

\item{...}{Passed to CGGPplotblocks}
}
\value{
Either makes plot or returns plot object
}
\description{
There are a few different plot functions for CGGP objects:
`CGGPplotblocks`, `CGGPplotblockselection`, 
`CGGPplotcorr`, `CGGPplotheat`, `CGGPplothist`,
`CGGPvalplot`, 
`CGGPplotslice`, `CGGPplotslice`, and `CGGPplotvariogram`.
Currently `CGGPplotblocks` is the default plot object.
}
\examples{
SG = CGGPcreate(3,100)
plot(SG)
}
