% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_S3_methods.R
\name{plot.SGGP}
\alias{plot.SGGP}
\title{S3 plot method for SGGP}
\usage{
\method{plot}{SGGP}(x, y, ...)
}
\arguments{
\item{x}{SGGP object}

\item{y}{Don't use}

\item{...}{Passed to SGGPblockplot}
}
\value{
Either makes plot or returns plot object
}
\description{
There are a few different plot functions for SGGP objects:
`SGGPblockplot`, `SGGPcorrplot`, `SGGPprojectionplot`,
`SGGPvalplot`, `SGGPheat`, and `SGGPhist`.
Currently `SGGPblockplot` is the default plot object.
}
\examples{
SG = SGGPcreate(3,100)
plot(SG)
}
