% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGGP_S3_methods.R
\name{print.CGGP}
\alias{print.CGGP}
\title{Print CGGP object}
\usage{
\method{print}{CGGP}(x, ...)
}
\arguments{
\item{x}{CGGP object}

\item{...}{Passed to print}
}
\value{
String to be printed
}
\description{
Default print as a list is bad since there's a lot of elements.
}
\examples{
SG = CGGPcreate(3,21)
print(SG)
f <- function(x) {x[1]+exp(x[2]) + log(x[3]+4)}
y <- apply(SG$design, 1, f)
SG <- CGGPfit(SG, y)
print(SG)
}
