% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_fastmatclcr}
\alias{rcpp_fastmatclcr}
\title{rcpp_fastmatclcr}
\usage{
rcpp_fastmatclcr(I, w, MSEmat, S, maxlevel)
}
\arguments{
\item{I}{Matrix}

\item{w}{vector}

\item{MSEmat}{Matrix}

\item{S}{Vector}

\item{maxlevel}{Integer}
}
\value{
Nothing, void
}
\description{
rcpp_fastmatclcr
}
