% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_fastmatclcranddclcr}
\alias{rcpp_fastmatclcranddclcr}
\title{rcpp_fastmatclcranddclcr}
\usage{
rcpp_fastmatclcranddclcr(I, w, MSEmat, dMSEmat, S, dS, maxlevel, numpara)
}
\arguments{
\item{I}{Matrix}

\item{w}{vector}

\item{MSEmat}{Matrix}

\item{dMSEmat}{Matrix}

\item{S}{Vector}

\item{dS}{Matrix}

\item{maxlevel}{Integer}

\item{numpara}{Integer}
}
\value{
Nothing, void
}
\description{
rcpp_fastmatclcranddclcr
}
