% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_gkronDBS}
\alias{rcpp_gkronDBS}
\title{rcpp_kronDBS}
\usage{
rcpp_gkronDBS(A, dA, B, p)
}
\arguments{
\item{A}{Vector}

\item{dA}{Vector}

\item{B}{Vector}

\item{p}{Vector}
}
\value{
kronDBS calculation
}
\description{
rcpp_kronDBS
}
\examples{
rcpp_gkronDBS(c(1,1), c(0,0), c(.75), c(1,1))
}
