% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpparma_outerproduct}
\alias{rcpparma_outerproduct}
\title{Outer}
\usage{
rcpparma_outerproduct(x)
}
\arguments{
\item{x}{vec}
}
\value{
outer prod
}
\description{
description
}
\examples{
rcpparma_outerproduct(c(1,0,3,7))
}
