% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGcreate.R
\name{SGcreate}
\alias{SGcreate}
\title{Create sparse grid GP}
\usage{
SGcreate(d, batchsize, corr = "Matern32", nugget = 0, xmin = rep(0,
  d), xmax = rep(1, d))
}
\arguments{
\item{d}{Input dimension}

\item{batchsize}{Number added to design each batch}

\item{corr}{Correlation function to use.}

\item{nugget}{Nugget term added to diagonal of correlation matrix,
for now only on predictions}

\item{xmin}{Min x values, vector. Must be rep(0,d).}

\item{xmax}{Max x values, vector. Must be rep(1,d).}
}
\value{
SGGP
}
\description{
Create sparse grid GP
}
\examples{
d <- 8
SG = SGcreate(d,201)
}
