\name{SGGP-package}
\alias{SGGP-package}
\alias{SGGP}
\docType{package}
\title{
\packageTitle{SGGP}
}
\description{
\packageDescription{SGGP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SGGP}
\packageIndices{SGGP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SGGP}

Maintainer: \packageMaintainer{SGGP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
