% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP.R
\docType{package}
\name{SGGP}
\alias{SGGP}
\alias{SGGP-package}
\title{SGGP: A package for running sparse grid computer experiments}
\description{
The SGGP package implements the method presented in Plumlee et al. (2019).
}
\section{SGGP functions}{

The SGGP functions: SGGPcreate, SGGPfit, SGGPappend, and SGGPpred
}

