% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_append_fs.R
\name{SGGP_internal_calcMSE}
\alias{SGGP_internal_calcMSE}
\title{Calculate MSE over single dimension for Matern 3/2}
\usage{
SGGP_internal_calcMSE(xl, theta, CorrMat)
}
\arguments{
\item{xl}{Vector of points in 1D}

\item{theta}{Correlation parameters}

\item{CorrMat}{Function that gives correlation matrix for vectors of 1D points.}
}
\value{
MSE value
}
\description{
Calcaluted using grid of integration points.
Can be calculated exactly, but not much reason in 1D.
}
\examples{
SGGP_internal_calcMSE(xl=c(0,.5,.9), theta=1,
         CorrMat=CorrMatMatern32)
}
