% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fastcalcassist_fs.R
\name{SGGP_internal_calcpw}
\alias{SGGP_internal_calcpw}
\title{Calculate predictive weights for SGGP}
\usage{
SGGP_internal_calcpw(SGGP, y, theta, return_lS = FALSE)
}
\arguments{
\item{y}{Measured values for SGGP$design}

\item{theta}{Correlation parameters}

\item{return_lS}{Should lS be returned?}

\item{SG}{SGGP object}
}
\value{
Vector with predictive weights
}
\description{
Predictive weights are Sigma^{-1}*y in standard GP.
This calculation is much faster since we don't need to
solve the full system of equations.
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SGGP$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
calculate_pw_C(SG=SG, y=y, logtheta=c(-.1,.1,.3))
}
