% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_corr_fs.R
\name{SGGP_internal_CorrMatCauchySQT}
\alias{SGGP_internal_CorrMatCauchySQT}
\title{Calculate correlation matrix for two sets of points in one dimension}
\usage{
SGGP_internal_CorrMatCauchySQT(x1, x2, theta, ...,
  return_dCdtheta = FALSE, return_numpara = FALSE)
}
\value{
Matrix
}
\description{
Note that this is not the correlation between two vectors.
It is for two
}
\examples{
CorrMatMatern32(c(0,.2,.4),c(.1,.3,.5), theta=.1)
}
