% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_corr_fs.R
\name{SGGP_internal_CorrMatPowerExp}
\alias{SGGP_internal_CorrMatPowerExp}
\title{Power exponential correlation function}
\usage{
SGGP_internal_CorrMatPowerExp(x1, x2, theta, return_dCdtheta = FALSE,
  return_numpara = FALSE)
}
\arguments{
\item{x1}{Vector of coordinates from same dimension}

\item{x2}{Vector of coordinates from same dimension}

\item{theta}{Correlation parameters:
\itemize{
    \item LS Log of parameter that controls lengthscale
    \item FD Logit of 0.5*parameter  that controls the fractal demension
    \item HE Log of parameter that controls the hurst effect
}}

\item{return_dCdtheta}{Should dCdtheta be returned?}

\item{return_numpara}{Should it just return the number of parameters?}
}
\value{
Matrix of correlation values between x1 and x2
}
\description{
Calculate correlation matrix for two sets of points in one dimension.
Note that this is not the correlation between two vectors.
}
\examples{
SGGP_internal_CorrMatPowerExp(c(0,.2,.4),c(.1,.3,.5), theta=c(-.7,.2))
}
\seealso{
Other correlation functions: \code{\link{SGGP_internal_CorrMatCauchySQT}},
  \code{\link{SGGP_internal_CorrMatCauchySQ}},
  \code{\link{SGGP_internal_CorrMatCauchy}},
  \code{\link{SGGP_internal_CorrMatGaussian}},
  \code{\link{SGGP_internal_CorrMatMatern32}},
  \code{\link{SGGP_internal_CorrMatMatern52}}
}
\concept{correlation functions}
