% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{SGGP_internal_gneglogpost}
\alias{SGGP_internal_gneglogpost}
\title{Gradient of likelihood. Is it log likelihood?}
\usage{
SGGP_internal_gneglogpost(theta, SGGP, y, return_lik = FALSE)
}
\arguments{
\item{theta}{Log of correlation parameters}

\item{SGGP}{SGGP object}

\item{y}{SGGP$design measured values}

\item{return_lik}{If yes, it returns a list with lik and glik}
}
\value{
Vector for gradient of likelihood w.r.t. x (theta)
}
\description{
Gradient of likelihood. Is it log likelihood?
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2})
SGGP_internal_gneglogpost(SG$thetaMAP, SG=SG, y=y)
}
