% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_pred_fs.R
\name{SGGP_internal_MSEpredcalc}
\alias{SGGP_internal_MSEpredcalc}
\title{Calculate MSE over one dimension}
\usage{
SGGP_internal_MSEpredcalc(xp, xl, theta, CorrMat)
}
\arguments{
\item{xp}{Points at which to calculate MSE}

\item{xl}{Levels along dimension, vector???}

\item{theta}{Correlation parameters}

\item{CorrMat}{Function that gives correlation matrix for vectors of 1D points.
for vector of 1D points.}
}
\value{
MSE predictions
}
\description{
Calculate MSE over one dimension
}
\examples{
MSEpred_calc(c(.4,.52), c(0,.25,.5,.75,1), theta=.1, nugget=1e-5,
             CorrMat=CorrMatMatern32,
             diag_corrMat=diag_corrMatMatern32)
}
