% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{SGGP_internal_neglogpost}
\alias{SGGP_internal_neglogpost}
\title{Calculate negative log posterior}
\usage{
SGGP_internal_neglogpost(theta, SGGP, y)
}
\arguments{
\item{theta}{Correlation parameters}

\item{SGGP}{SGGP object}

\item{y}{Measured values of SGGP$design}
}
\value{
Likelihood
}
\description{
Calculate negative log posterior
}
\examples{
SG <- SGGPcreate(d=3, batchsize=20)
Y <- apply(SG$design, 1, function(x){x[1]+x[2]^2})
SG <- SGGPfit(SG, Y)
SGGP_internal_neglogpost(SG$thetaMAP, SG=SG, y=SG$y)
}
