% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_pred_fs.R
\name{SGGP_internal_postvarmatcalcfasterasym6}
\alias{SGGP_internal_postvarmatcalcfasterasym6}
\title{Calculate posterior variance, faster version}
\usage{
SGGP_internal_postvarmatcalcfasterasym6(GMat1, GMat2, cholS, INDSN)
}
\arguments{
\item{GMat1}{Matrix 1}

\item{GMat2}{Matrix 2}

\item{cholS}{Cholesky factorization of S}

\item{INDSN}{Indices, maybe}
}
\value{
Variance posterior
}
\description{
Calculate posterior variance, faster version
}
\examples{
SGGP_internal_postvarmatcalc(c(.4,.52), c(0,.25,.5,.75,1),
             xo=c(.11), theta=c(.1,.2,.3),
             CorrMat=SGGP_internal_CorrMatCauchySQT)
}
