% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_useful.R
\name{SGGP_internal_set_corr}
\alias{SGGP_internal_set_corr}
\title{Set correlation function of SGGP object}
\usage{
SGGP_internal_set_corr(SGGP, corr)
}
\arguments{
\item{SGGP}{SGGP object}

\item{corr}{Correlation function}
}
\value{
SGGP object
}
\description{
Set correlation function of SGGP object
}
\examples{
obj <- SGGPcreate(3, 20, corr="matern52")
SGGP_internal_set_corr(obj, "gaussian")
}
