% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_append_fs.R
\name{SGGPappend}
\alias{SGGPappend}
\title{Add points to SGGP}
\usage{
SGGPappend(SGGP, batchsize, selectionmethod = "UCB")
}
\arguments{
\item{SGGP}{Sparse grid object}

\item{batchsize}{Number of points to add}

\item{selectionmethod}{How points will be selected: one of `UCB`, `TS`, and `Greedy`}
}
\value{
SG with new points added.
}
\description{
Add `batchsize` points to `SG` using `theta`.
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
SG <- SGappend(theta=c(.1,.1,.1), SG=SG, batchsize=20)
# UCB,TS,Greedy
}
