% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPblockplot}
\alias{SGGPblockplot}
\title{SGGP block plot}
\usage{
SGGPblockplot(SGGP, singleplot = TRUE)
}
\arguments{
\item{SGGP}{SGGP object}

\item{singleplot}{If only two dimensions, should a single plot be made?}
}
\value{
ggplot2 plot
}
\description{
Plot the 2D projections of the blocks of an SGGP object.
}
\examples{
# The first and fourth dimensions are most active and will have greater depth
ss <- SGGPcreate(d=5, batchsize=50)
f <- function(x) {cos(2*pi*x[1]*3) + x[3]*exp(4*x[4])}
ss <- SGGPfit(ss, Y=apply(ss$design, 1, f))
ss <- SGGPappend(SGGP=ss, batchsize=100)
SGGPblockplot(ss)

mat <- matrix(c(1,1,1,2,2,1,2,2,1,3), ncol=2, byrow=TRUE)
SGGPblockplot(mat)
}
