% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_create_fs.R
\name{SGGPcreate}
\alias{SGGPcreate}
\title{Create sparse grid GP}
\usage{
SGGPcreate(d, batchsize)
}
\arguments{
\item{d}{Input dimension}

\item{batchsize}{Number added to design each batch}

\item{xmin}{Min x values, vector. Must be rep(0,d).}

\item{xmax}{Max x values, vector. Must be rep(1,d).}

\item{nugget}{Nugget term added to diagonal of correlation matrix,
for now only on predictions}

\item{corr}{Correlation function to use.}
}
\value{
SGGP
}
\description{
Create sparse grid GP
}
\examples{
d <- 8
SG = SGcreate(d,201)
}
