% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{SGGPfit}
\alias{SGGPfit}
\title{Update SGGP model given data}
\usage{
SGGPfit(SGGP, Y, ..., Xs = NULL, Ys = NULL, theta0 = SGGP$thetaMAP,
  laplaceapprox = TRUE, HandlingSuppData = SGGP$HandlingSuppData,
  lower = rep(-1, SGGP$numpara * SGGP$d), upper = rep(1, SGGP$numpara *
  SGGP$d), use_PCA = SGGP$use_PCA,
  separateoutputparameterdimensions = is.matrix(SGGP$thetaMAP),
  use_progress_bar = TRUE, Ynew)
}
\arguments{
\item{SGGP}{Sparse grid objects}

\item{Y}{Output values calculated at SGGP$design}

\item{...}{Forces you to name arguments.}

\item{Xs}{Supplemental X matrix}

\item{Ys}{Supplemental Y values}

\item{theta0}{Initial theta}

\item{laplaceapprox}{Should Laplace approximation be used?}

\item{HandlingSuppData}{How should supplementary data be handled?
* Correct: full likelihood with grid and supplemental data
* Only: only use supplemental data
* Ignore: ignore supplemental data
* Mixture: sum of grid LLH and supplemental LLH, not statistically valid
* MarginalValidation: a validation shortcut
* FullValidation: a validation shortcut}

\item{lower}{Lower bound for parameter optimization}

\item{upper}{Upper bound for parameter optimization}

\item{use_PCA}{Should PCA be used if output is multivariate?}

\item{separateoutputparameterdimensions}{If multiple output dimensions,
should separate parameters be fit to each dimension?}

\item{use_progress_bar}{If using MCMC sampling, should a progress bar be
displayed?}

\item{Ynew}{Values of `SGGP$design_unevaluated`}
}
\value{
Updated SGGP object fit to data given
}
\description{
Update SGGP model given data
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2})
SG <- SGGPfit(SG=SG, Y=y)
}
\seealso{
Other SGGP core functions: \code{\link{SGGPappend}},
  \code{\link{SGGPcreate}}, \code{\link{predict.SGGP}}
}
\concept{SGGP core functions}
