% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPheat}
\alias{SGGPheat}
\title{Heatmap of SG design depth}
\usage{
SGGPheat(SGGP)
}
\arguments{
\item{SGGP}{SGGP object}
}
\value{
A heat map made from ggplot2
}
\description{
The values on the diagonal are largest design depth for that dimension.
The off-diagonal values are the largest design depth that both dimensions
have been measured at simultaneously.
A greater depth means that more points have been measured along that
dimension or two-dimensional subspace.
}
\examples{
\dontrun{
# All dimensions should look similar
d <- 8
SG = SGGPcreate(d,201)
SGGPheat(SG)

# The first and fourth dimensions are most active and will have greater depth
SG <- SGGPcreate(d=5, batchsize=50)
f <- function(x) {cos(2*pi*x[1]*3) + exp(4*x[4])}
for (i in 1:1) {
  SG <- SGGPfit(SG, Y=apply(SG$design, 1, f))
  SG <- SGGPappend(SGGP=SG, batchsize=200)
}
# SG <- SGGPfit(SG, Y=apply(SG$design, 1, f))
SGGPheat(SG)
}
}
\references{
https://stackoverflow.com/questions/14290364/heatmap-with-values-ggplot2
}
