% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPplotprojection}
\alias{SGGPplotprojection}
\title{SGGP projection plot}
\usage{
SGGPplotprojection(SGGP, proj = 0.5, np = 300, color = "pink",
  outdims)
}
\arguments{
\item{SGGP}{SGGP object}

\item{proj}{Point to project onto}

\item{np}{Number of points to use along each dimension}

\item{color}{Color to make error region}

\item{outdims}{If multiple outputs, which of them should be plotted?}
}
\value{
ggplot2 object
}
\description{
Show prediction plots when projected down to one dimension.
Most useful when setting all values to 0.5 because it will
have the most points.
}
\examples{
\dontrun{
d <- 5
f1 <- function(x){x[1]+x[2]^2 + cos(x[3]^2*2*pi*4) - 3.3}
s1 <- SGGPcreate(d, 200)
s1 <- SGGPfit(s1, apply(s1$design, 1, f1))
#s1 <- SGGPappend(s1, 200)
#s1 <- SGGPfit(s1, apply(s1$design, 1, f1))
SGGPplotprojection(s1)
SGGPplotprojection(s1, 0.)
SGGPplotprojection(s1, s1$design[nrow(s1$design),])
}
}
