% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPplotsamplesneglogpost}
\alias{SGGPplotsamplesneglogpost}
\title{Plot negative log posterior likelihood of samples}
\usage{
SGGPplotsamplesneglogpost(SGGP)
}
\arguments{
\item{SGGP}{SGGP object}
}
\value{
ggplot2 object
}
\description{
Plot negative log posterior likelihood of samples
}
\examples{
gs <- SGGPcreate(d=3, batchsize=100)
f <- function(x){x[1]^1.2+x[3]^.4*sin(2*pi*x[2]^2*3) + .1*exp(3*x[3])}
y <- apply(gs$design, 1, f)
gs <- SGGPfit(gs, Y=y)
SGGPplotsamplesneglogpost(gs)
}
