% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_S3_methods.R, R/SGGPpredstuff.R
\name{predict.SGGP}
\alias{predict.SGGP}
\alias{SGGPpred}
\title{S3 predict method for SGGP}
\usage{
\method{predict}{SGGP}(object, xp, y, logtheta, theta, ...)

SGGPpred(xp, SG, y, ..., logtheta = SG$logtheta, theta)
}
\arguments{
\item{xp}{x value to predict at}

\item{y}{Observations for SG}

\item{logtheta}{Log of correlation parameters}

\item{theta}{Correlation parameters}

\item{...}{Don't use, just forces theta to be named}

\item{SG}{SG object}
}
\value{
Predicted mean values
}
\description{
Passes to SGGPpred

Predict using SG with y values at xp?
Shouldn't y values already be stored in SG?
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
SGGPpred(matrix(c(.1,.1,.1),1,3), SG=SG, y=y, theta=c(.1,.1,.1))
cbind(SGGPpred(SG$design, SG=SG, y=y, theta=c(.1,.1,.1))$mean, y) # Should be near equal
}
