% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPpredstuff.R
\name{SGGPpredMV}
\alias{SGGPpredMV}
\title{Multivariate SGGP prediction}
\usage{
SGGPpredMV(xp, SG, yMV, ..., logtheta, theta)
}
\arguments{
\item{xp}{x value to predict at}

\item{SG}{SG object}

\item{yMV}{Output matrix, each row is for a design point}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta}{Log of correlation parameters}

\item{theta}{Correlation parameters}
}
\value{
Matrix of predictions
}
\description{
Predict output for an SGGP with multivariate output.
}
\examples{

SG <- SGcreate(c(0,0,0), c(1,1,1), batchsize=100)
y1 <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
y2 <- apply(SG$design, 1, function(x){x[1]^1.3+.4*sin(6*x[2])+rnorm(1,0,.01)})
y <- cbind(y1, y2)
tx <- cbind(SGGPpredMV(SG$design, SG=SG, yMV=y, theta=c(.1,.1,.1))$mean, y)
tx # Columns 1 and 3, and 2 and 4 should be near equal
cor(tx)
}
