% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPvalplot}
\alias{SGGPvalplot}
\title{Plot validation prediction errors}
\usage{
SGGPvalplot(SG, y, xval, yval, ypred)
}
\arguments{
\item{SG}{SGGP object}

\item{y}{Measurements at SG$design}

\item{ypred}{(optional) Predictions of SG at Xval}

\item{Xval}{X validation data}

\item{Yval}{Y validation data}
}
\value{
None, makes a plot
}
\description{
Plot validation prediction errors
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
f1 <- function(x){x[1]+x[2]^2+rnorm(1,0,.01)}
y <- apply(SG$design, 1, f1)
Xval <- matrix(runif(3*100), ncol=3)
Yval <- apply(Xval, 1, f1)
SG <- logthetaVALID(SG=SG, y=y, xval=Xval, yval=Yval)
SGGPvalplot(SG=SG, y=y, xval=Xval, yval=Yval)
}
