% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPvalstats}
\alias{SGGPvalstats}
\title{Calculate stats for SGGP prediction on validation data}
\usage{
SGGPvalstats(SGGP, Xval, Yval, d = NULL)
}
\arguments{
\item{SGGP}{SGGP object}

\item{Xval}{X validation matrix}

\item{Yval}{Y validation data}

\item{d}{Dimension if multivariate output}
}
\value{
data frame
}
\description{
Calculate stats for SGGP prediction on validation data
}
\examples{
SGGPvalstats
}
