% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{SGheat}
\alias{SGheat}
\title{Heatmap of SG design depth}
\usage{
SGheat(SG)
}
\arguments{
\item{SG}{SGGP object}
}
\value{
A heat map made from ggplot2
}
\description{
The values on the diagonal are largest design depth for that dimension.
The off-diagonal values are the largest design depth that both dimensions
have been measured at simultaneously.
A greater depth means that more points have been measured along that
dimension or two-dimensional subspace.
}
\examples{
# All dimensions should look similar
d <- 8
SG = SGcreate(d,201)
SGheat(SG)

# The first dimension is more active and will have greater depth
SG <- SGcreate(d=5, batchsize=10)
SG <- SGappend(logtheta=c(-2,2,2,2,2), SG=SG, batchsize=100)
SGheat(SG)
}
\references{
https://stackoverflow.com/questions/14290364/heatmap-with-values-ggplot2
}
