% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{SGhist}
\alias{SGhist}
\title{Histogram of measurements at each design depth of each input dimension}
\usage{
SGhist(SG, ylog = T)
}
\arguments{
\item{SG}{SGGP object}

\item{ylog}{Should the y axis be put on a log scale?}
}
\value{
Histogram plot made using ggplot2
}
\description{
A greater design depth signifies a more important dimension.
Thus a larger right tail on the histogram are more important variables.
}
\examples{
# All dimensions should look similar
d <- 8
SG = SGcreate(d,201)
SGhist(SG)
SGhist(SG, ylog=F)

# The first dimension is more active and will have greater depth
SG <- SGcreate(d=5, batchsize=10)
SG <- SGappend(logtheta=c(-2,2,2,2,2), SG=SG, batchsize=100)
SGhist(SG)
}
