% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{updateSG}
\alias{updateSG}
\title{Update SGGP object with new data}
\usage{
updateSG(SG, y, restarts = 4, ..., ynew, method = "MLE")
}
\arguments{
\item{SG}{SGGP object}

\item{y}{Measured output values}

\item{restarts}{Number of optimization restarts}

\item{...}{Not used}

\item{ynew}{If only returning new output values from last append, use this}

\item{method}{Whether parameters should be updated using MLE or validation.}
}
\value{
Updated SG object
}
\description{
Update SGGP object with new data
}
\examples{
f <- function(x) {(x[1]-.5)^2 + exp(-x[2]) + sin(2*pi*x[3])}
SG <- SGcreate(d=3, batchsize=100)
SG <- updateSG(SG, apply(SG$design, 1, f))
#SG <- SGappend(SG=SG, batchsize=20)
}
