% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{valstats}
\alias{valstats}
\title{Calculate stats for prediction on validation data}
\usage{
valstats(predmean, predvar, Yval, bydim = TRUE)
}
\arguments{
\item{predmean}{Predicted mean}

\item{predvar}{Predicted variance}

\item{Yval}{Y validation data}

\item{bydim}{If multiple outputs, should it be done separately by dimension?}
}
\value{
data frame
}
\description{
Calculate stats for prediction on validation data
}
\examples{
valstats(c(0,1,2), c(.01,.01,.01), c(0,1.1,1.9))
valstats(cbind(c(0,1,2), c(1,2,3)),
         cbind(c(.01,.01,.01),c(.1,.1,.1)),
         cbind(c(0,1.1,1.9),c(1,2,3)))
valstats(cbind(c(0,1,2), c(8,12,34)),
         cbind(c(.01,.01,.01),c(1.1,.81,1.1)),
         cbind(c(0,1.1,1.9),c(10,20,30)), bydim=FALSE)
valstats(cbind(c(.8,1.2,3.4), c(8,12,34)),
         cbind(c(.01,.01,.01),c(1.1,.81,1.1)),
         cbind(c(1,2,3),c(10,20,30)), bydim=FALSE)
}
\references{
Gneiting, Tilmann, and Adrian E. Raftery.
"Strictly proper scoring rules, prediction, and estimation."
Journal of the American Statistical Association 102.477 (2007): 359-378.
}
