% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPstuff.R
\name{dCorrMat}
\alias{dCorrMat}
\title{Calculate derivative of CorrMat with respect to theta}
\usage{
dCorrMat(x1, x2, ..., logtheta, theta)
}
\arguments{
\item{x1}{First vector of 1D points}

\item{x2}{Second vector of 1D points}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta}{Log correlation parameter}

\item{theta}{Correlation parameter}
}
\value{
Matrix
}
\description{
Calculate derivative of CorrMat with respect to theta
}
\examples{
dCorrMat(c(0,.2,.4),c(.1,.3,.5), theta=.1)
}
