% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPlik.R
\name{logthetaMLE}
\alias{logthetaMLE}
\title{Calculate theta MLE given data}
\usage{
logthetaMLE(SG, y, ..., logtheta0 = rep(0, SG$d), tol = 0.1)
}
\arguments{
\item{SG}{Sparse grid objects}

\item{y}{Output values calculated at SG$design}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta0}{Initial logtheta}

\item{tol}{Tolerance for optimization}
}
\value{
theta MLE
}
\description{
Calculate theta MLE given data
}
\examples{
SG <- SGcreate(c(0,0,0), c(1,1,1), batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
logthetaMLE(SG=SG, y=y)
}
