% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPlik.R
\name{logthetaMLE}
\alias{logthetaMLE}
\title{Calculate theta MLE given data}
\usage{
logthetaMLE(SG, y, ..., lower = rep(-2, SG$d), upper = rep(3.9, SG$d),
  method = "L-BFGS-B", logtheta0 = rep(0, SG$d), tol = 1e-04,
  return_optim = FALSE, use_splitfngr = FALSE)
}
\arguments{
\item{SG}{Sparse grid objects}

\item{y}{Output values calculated at SG$design}

\item{...}{Don't use, just forces theta to be named}

\item{lower}{Lower bound for parameter optimization}

\item{upper}{Upper bound for parameter optimization}

\item{method}{Optimization method, must be "L-BFGS-B" when using lower and upper}

\item{logtheta0}{Initial logtheta}

\item{tol}{Relative tolerance for optimization. Can't use absolute tolerance
since lik can be less than zero.}

\item{return_optim}{If TRUE, return output from optim().}

\item{use_splitfngr}{Should give exact same results but with a slight speed up
If FALSE return updated SG.}
}
\value{
theta MLE
}
\description{
Calculate theta MLE given data
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
logthetaMLE(SG=SG, y=y)
}
