% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPlik.R
\name{logthetaMLEMV}
\alias{logthetaMLEMV}
\title{Find MLE of logtheta for multivariate output}
\usage{
logthetaMLEMV(SG, yMV, ..., logtheta0 = rep(0, SG$d), tol = 1e-04,
  return_optim)
}
\arguments{
\item{SG}{SG object}

\item{yMV}{Output matrix}

\item{...}{Don't use}

\item{logtheta0}{Initial values of logtheta for optimization}

\item{tol}{Optimization tolerance}

\item{return_optim}{If TRUE, return output from optim().
If FALSE return updated SG.}
}
\value{
Vector, logtheta MLE
}
\description{
Find MLE of logtheta for multivariate output
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y1 <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
y2 <- apply(SG$design, 1, function(x){x[1]^1.3+.4*sin(6*x[2])+rnorm(1,0,.01)})
y <- cbind(y1, y2)
logthetaMLE(SG=SG, y=y1)
logthetaMLE(SG=SG, y=y2)
logthetaMLEMV(SG=SG, yMV=y)
}
