% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPappendstuff.R
\name{MSE_calc}
\alias{MSE_calc}
\title{Calculate MSE over single dimension for Matern 3/2}
\usage{
MSE_calc(xl, ..., logtheta, theta)
}
\arguments{
\item{xl}{Vector of points in 1D}

\item{theta}{Correlation parameters}
}
\value{
MSE value
}
\description{
Equation found using
}
\examples{
MSE_calc(xl=c(0,.5,.9), theta=-1)
}
