% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPpredstuff.R
\name{MSEpred_calc}
\alias{MSEpred_calc}
\title{????????????}
\usage{
MSEpred_calc(xp, xl, ..., logtheta, theta)
}
\arguments{
\item{xp}{Points at which to calculate MSE}

\item{xl}{Levels along dimension, vector???}

\item{theta}{Correlation parameters on log scale}
}
\value{
MSE predictions
}
\description{
????????????
}
\examples{
MSEpred_calc(c(.4,.52), c(0,.25,.5,.75,1), th=.1)
}
