% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPpredstuff.R
\name{MSEpred_calc}
\alias{MSEpred_calc}
\title{????????????}
\usage{
MSEpred_calc(xp, xl, ..., logtheta, theta, nugget)
}
\arguments{
\item{xp}{Points at which to calculate MSE}

\item{xl}{Levels along dimension, vector???}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta}{Log of correlation parameters}

\item{theta}{Correlation parameters}

\item{nugget}{Nugget to add to diagonal of correlation matrix}
}
\value{
MSE predictions
}
\description{
????????????
}
\examples{
MSEpred_calc(c(.4,.52), c(0,.25,.5,.75,1), theta=.1, nugget=1e-5)
}
