% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_S3_methods.R
\name{print.SGGP}
\alias{print.SGGP}
\title{Print SGGP object}
\usage{
\method{print}{SGGP}(x, ...)
}
\arguments{
\item{x}{SGGP object}

\item{...}{Passed to print}
}
\value{
String to be printed
}
\description{
Default print as a list is bad since there's a lot of elements.
}
\examples{
SG = SGcreate(3,201)
print(SG)
}
