% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_kronDBS}
\alias{rcpp_kronDBS}
\title{rcpp_kronDBS}
\usage{
rcpp_kronDBS(A, B, p)
}
\arguments{
\item{A}{Vector}

\item{B}{Vector}

\item{p}{Vector}
}
\value{
kronDBS calculation
}
\description{
rcpp_kronDBS
}
