% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPappendstuff.R
\name{SGappend}
\alias{SGappend}
\title{Add points to SGGP}
\usage{
SGappend(SG, batchsize, ..., logtheta, theta)
}
\arguments{
\item{SG}{Sparse grid object}

\item{batchsize}{Number of points to add}

\item{...}{Don't use, just forces theta to be named}

\item{logtheta}{Log of theta, give one of theta and logtheta}

\item{theta}{Correlation parameters}
}
\value{
SG with new points added.
}
\description{
Add `batchsize` points to `SG` using `theta`.
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
SG <- SGappend(theta=c(.1,.1,.1), SG=SG, batchsize=20)
}
