% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGcreate.R
\name{SGcreate}
\alias{SGcreate}
\title{Create sparse grid GP}
\usage{
SGcreate(xmin, xmax, batchsize, nugget = 0)
}
\arguments{
\item{xmin}{Min x values, vector}

\item{xmax}{Max x values, vector}

\item{batchsize}{Number added to design each batch}

\item{nugget}{Nugget term added to diagonal of correlation matrix, for now only on predictions}
}
\value{
SGGP
}
\description{
Create sparse grid GP
}
\examples{
d <- 8
SG = SGcreate(rep(0, d), rep(1, d),201)
}
