% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_append_fs.R
\name{SGGP_internal_calcMSEde}
\alias{SGGP_internal_calcMSEde}
\title{Calculate MSE over blocks}
\usage{
SGGP_internal_calcMSEde(valsinds, MSE_MAP)
}
\arguments{
\item{valsinds}{Block levels to calculate MSEs for}

\item{MSE_MAP}{Matrix of MSE values}
}
\value{
All MSE values
}
\description{
Delta of adding block is product over i=1..d of IMSE(i,j-1) - IMSE(i,j)
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
theta <- c(.1,.1,.1)
MSE_MAP <- outer(1:SGGP$d, 1:8, 
     Vectorize(function(lcv1, dimlcv) {
        MSE_calc(SGGP$xb[1:SGGP$sizest[dimlcv]], theta=theta[lcv1], nugget=0,
         CorrMat=CorrMatMatern32,
         diag_corrMat=diag_corrMatMatern32)
 }))
MSE_de(SGGP$po[1:SGGP$poCOUNT, ], MSE_MAP)
}
