% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_corr_fs.R
\name{SGGP_internal_CorrMatCauchySQT}
\alias{SGGP_internal_CorrMatCauchySQT}
\title{Calculate correlation matrix for two sets of points in one dimension}
\usage{
SGGP_internal_CorrMatCauchySQT(x1, x2, theta, return_dCdtheta = FALSE,
  return_numpara = FALSE)
}
\arguments{
\item{x1}{Vector of coordinates from same dimension}

\item{x2}{Vector of coordinates from same dimension}

\item{theta}{Correlation parameters:
\itemize{
    \item LS Log of parameter that controls lengthscale
    \item FD Logit of 0.5*parameter  that controls the fractal demension
    \item HE Log of parameter that controls the hurst effect
}}

\item{return_dCdtheta}{Should dCdtheta be returned?}

\item{return_numpara}{Should it just return the number of parameters?}
}
\value{
Matrix
}
\description{
Note that this is not the correlation between two vectors.
}
\examples{
SGGP_internal_CorrMatCauchySQT(c(0,.2,.4),c(.1,.3,.5), theta=c(-.1,.3,-.7))
}
