% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGP_internal_CorrPlot}
\alias{SGGP_internal_CorrPlot}
\title{Plot correlation samples}
\usage{
SGGP_internal_CorrPlot(Corr = SGGP_internal_CorrMatGaussian,
  theta = NULL, numlines = 20, plot_with = "ggplot", zero = TRUE)
}
\arguments{
\item{Corr}{Correlation function or SGGP object.
If SGGP object, it will make plots for thetaMAP,
the max a posteriori theta.}

\item{theta}{Parameters for Corr}

\item{numlines}{Number of sample paths to draw}

\item{plot_with}{Should "base" or "ggplot2" be used to make the plot?}

\item{zero}{Should the sample paths start at y=0?}
}
\value{
Plot
}
\description{
Plot samples for a given correlation function and parameters.
Useful for getting an idea of what the correlation parameters mean
in terms of smoothness.
}
\examples{
SGGP_internal_CorrPlot()
SGGP_internal_CorrPlot(theta=c(-2,-1,0,1))

SG <- SGGPcreate(d=3, batchsize=100)
f <- function(x){x[1]^1.2+sin(2*pi*x[2]*3)}
y <- apply(SG$design, 1, f)
SG <- SGGPfit(SG, Y=y)
SGGP_internal_CorrPlot(SG)
}
