% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{SGGP_internal_gneglogpost}
\alias{SGGP_internal_gneglogpost}
\title{Gradient of negative log likelihood posterior}
\usage{
SGGP_internal_gneglogpost(theta, SGGP, y, ..., return_lik = FALSE,
  ys = NULL, Xs = NULL, HandlingSuppData = "Correct")
}
\arguments{
\item{theta}{Log of correlation parameters}

\item{SGGP}{SGGP object}

\item{y}{SGGP$design measured values}

\item{...}{Just a placeholder}

\item{return_lik}{If yes, it returns a list with lik and glik}

\item{ys}{Supplementary output data}

\item{Xs}{Supplementary input data}

\item{HandlingSuppData}{How should supplementary data be handled?}
}
\value{
Vector for gradient of likelihood w.r.t. x (theta)
}
\description{
Gradient of negative log likelihood posterior
}
\examples{
SG <- SGGPcreate(d=3, batchsize=20)
Y <- apply(SG$design, 1, function(x){x[1]+x[2]^2})
SG <- SGGPfit(SG, Y)
SGGP_internal_gneglogpost(SG$thetaMAP, SG=SG, y=SG$y)
}
