% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{SGGP_internal_neglogpost}
\alias{SGGP_internal_neglogpost}
\title{Calculate negative posterior}
\usage{
SGGP_internal_neglogpost(theta, SGGP, y)
}
\arguments{
\item{theta}{Correlation parameters}

\item{SGGP}{SGGP object}

\item{y}{Measured values of SGGP$design}
}
\value{
Likelihood
}
\description{
Calculate negative posterior
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SGGP$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
SGGP_internal_neglogpost(c(.1,.1,.1), SG=SG, y=y)
}
