% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_speedup_neglogpost.R
\name{SGGP_internal_neglogpost}
\alias{SGGP_internal_neglogpost}
\title{Calculate negative log posterior}
\usage{
SGGP_internal_neglogpost(theta, SGGP, y, ..., ys = NULL, Xs = NULL,
  HandlingSuppData = "Correct")
}
\arguments{
\item{theta}{Correlation parameters}

\item{SGGP}{SGGP object}

\item{y}{Measured values of SGGP$design}

\item{...}{Just a placeholder}

\item{ys}{Supplementary output data}

\item{Xs}{Supplementary input data}

\item{HandlingSuppData}{How should supplementary data be handled?
\itemize{
\item Correct: full likelihood with grid and supplemental data
\item Only: only use supplemental data
\item Ignore: ignore supplemental data
}}
}
\value{
Likelihood
}
\description{
Calculate negative log posterior
}
\examples{
SG <- SGGPcreate(d=3, batchsize=20)
Y <- apply(SG$design, 1, function(x){x[1]+x[2]^2})
SG <- SGGPfit(SG, Y)
SGGP_internal_neglogpost(SG$thetaMAP, SG=SG, y=SG$y)
}
