% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{SGGP_internal_postvarmatcalc}
\alias{SGGP_internal_postvarmatcalc}
\title{Calculate posterior variance}
\usage{
SGGP_internal_postvarmatcalc(x1, x2, xo, theta, CorrMat, ...,
  returndPVMC = FALSE, returndiagonly = FALSE)
}
\arguments{
\item{x1}{Points at which to calculate MSE}

\item{x2}{Levels along dimension, vector???}

\item{xo}{No clue what this is}

\item{theta}{Correlation parameters}

\item{CorrMat}{Function that gives correlation matrix for vectors of 1D points.
for vector of 1D points.}

\item{...}{Placeholder}

\item{returndPVMC}{Should dPVMC be returned?}

\item{returndiagonly}{Should only the diagonal be returned?}
}
\value{
Variance posterior
}
\description{
Calculate posterior variance
}
\examples{
SGGP_internal_postvarmatcalc(c(.4,.52), c(0,.25,.5,.75,1),
             xo=c(.11), theta=c(.1,.2,.3),
             CorrMat=SGGP_internal_CorrMatCauchySQT)
}
