% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_append_fs.R
\name{SGGPappend}
\alias{SGGPappend}
\title{Add points to SGGP}
\usage{
SGGPappend(SGGP, batchsize, selectionmethod = "UCB",
  RIMSEperpoint = TRUE, multioutputdim_weights = 1)
}
\arguments{
\item{SGGP}{Sparse grid object}

\item{batchsize}{Number of points to add}

\item{selectionmethod}{How points will be selected: one of `UCB`, `TS`, or `Greedy`}

\item{RIMSEperpoint}{Should RIMSE per point be used?}

\item{multioutputdim_weights}{Weights for each output dimension.}
}
\value{
SG with new points added.
}
\description{
Add `batchsize` points to `SG` using `theta`.
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2})
SG <- SGGPfit(SG, Y=y)
SG <- SGGPappend(SGGP=SG, batchsize=20)
# UCB,TS,Greedy
}
