% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_create_fs.R
\name{SGGPcreate}
\alias{SGGPcreate}
\title{Create sparse grid GP}
\usage{
SGGPcreate(d, batchsize, corr = "CauchySQT", grid_sizes = c(1, 2, 4, 4,
  8, 12, 32))
}
\arguments{
\item{d}{Input dimension}

\item{batchsize}{Number added to design each batch
for now only on predictions}

\item{corr}{Name of correlation function to use. Must be one of "CauchySQT",
"CauchySQ", "Cauchy", "Gaussian", "PowerExp", "Matern32", "Matern52".}

\item{grid_sizes}{Size of grid refinements.}
}
\value{
SGGP
}
\description{
Create sparse grid GP
}
\examples{
d <- 8
SG = SGGPcreate(d,200)
}
