% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_create_fs.R
\name{SGGPcreate}
\alias{SGGPcreate}
\title{Create sparse grid GP}
\usage{
SGGPcreate(d, batchsize)
}
\arguments{
\item{d}{Input dimension}

\item{batchsize}{Number added to design each batch
for now only on predictions}
}
\value{
SGGP
}
\description{
Create sparse grid GP
}
\examples{
d <- 8
SG = SGcreate(d,201)
}
