% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_create_fs.R
\name{SGGPcreate}
\alias{SGGPcreate}
\title{Create sparse grid GP}
\usage{
SGGPcreate(d, batchsize, corr = "CauchySQ", ..., grid_sizes = c(1, 2,
  4, 4, 8, 12, 32), Xs = NULL, Ys = NULL, HandlingSuppData = if
  (is.null(Xs)) {     "Ignore" } else {     "Correct" },
  supp_args = list())
}
\arguments{
\item{d}{Input dimension}

\item{batchsize}{Number added to design each batch
for now only on predictions}

\item{corr}{Name of correlation function to use. Must be one of "CauchySQT",
"CauchySQ", "Cauchy", "Gaussian", "PowerExp", "Matern32", "Matern52".}

\item{...}{Force named arguments}

\item{grid_sizes}{Size of grid refinements.}

\item{Xs}{Supplemental X data}

\item{Ys}{Supplemental Y data}

\item{HandlingSuppData}{How should supplementary data be handled?
* Correct: full likelihood with grid and supplemental data
* Only: only use supplemental data
* Ignore: ignore supplemental data
* Mixture: sum of grid LLH and supplemental LLH, not statistically valid
* MarginalValidation: a validation shortcut
* FullValidation: a validation shortcut}

\item{supp_args}{Arguments used to fit if Xs and Ys are given}
}
\value{
SGGP
}
\description{
Create sparse grid GP
}
\examples{
d <- 8
SG = SGGPcreate(d,200)
}
\seealso{
Other SGGP core functions: \code{\link{SGGPappend}},
  \code{\link{SGGPfit}}, \code{\link{predict.SGGP}}
}
\concept{SGGP core functions}
