% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{SGGPfit}
\alias{SGGPfit}
\title{Calculate theta MLE given data}
\usage{
SGGPfit(SGGP, Y, Xs = NULL, Ys = NULL, theta0 = rep(0, SGGP$numpara *
  SGGP$d), laplaceapprox = TRUE, lower = rep(-1, SGGP$numpara *
  SGGP$d), upper = rep(1, SGGP$numpara * SGGP$d),
  use_PCA = SGGP$use_PCA, Ynew)
}
\arguments{
\item{SGGP}{Sparse grid objects}

\item{Y}{Output values calculated at SGGP$design}

\item{Xs}{Supplemental X matrix}

\item{Ys}{Supplemental Y values}

\item{theta0}{Initial theta}

\item{laplaceapprox}{Should Laplace approximation be used?}

\item{lower}{Lower bound for parameter optimization}

\item{upper}{Upper bound for parameter optimization}

\item{use_PCA}{Should PCA be used if output is multivariate?}

\item{Ynew}{Values of `SGGP$design_unevaluated`}
}
\value{
Updated SGGP object fit to data given
}
\description{
Calculate theta MLE given data
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2})
SG <- SGGPfit(SG=SG, Y=y)
}
