% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_fit_fs.R
\name{SGGPfit}
\alias{SGGPfit}
\title{Calculate theta MLE given data}
\usage{
SGGPfit(SGGP, Y, Xs = NULL, Ys = NULL, theta0 = rep(0, SGGP$numpara *
  SGGP$d), laplaceapprox = TRUE, lower = rep(-1, SGGP$numpara *
  SGGP$d), upper = rep(1, SGGP$numpara * SGGP$d))
}
\arguments{
\item{SGGP}{Sparse grid objects}

\item{Y}{Output values calculated at SGGP$design}

\item{Xs}{Supplemental X matrix}

\item{Ys}{Supplemental Y values}

\item{theta0}{Initial theta}

\item{laplaceapprox}{Should Laplace approximation be used?}

\item{lower}{Lower bound for parameter optimization}

\item{upper}{Upper bound for parameter optimization}
}
\value{
theta MLE
}
\description{
Calculate theta MLE given data
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SGGP$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
SGGPfit(SG=SG, Y=y)
}
