% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPhist}
\alias{SGGPhist}
\title{Histogram of measurements at each design depth of each input dimension}
\usage{
SGGPhist(SGGP, ylog = TRUE)
}
\arguments{
\item{SGGP}{SGGP object}

\item{ylog}{Should the y axis be put on a log scale?}
}
\value{
Histogram plot made using ggplot2
}
\description{
A greater design depth signifies a more important dimension.
Thus a larger right tail on the histogram are more important variables.
}
\examples{
\dontrun{
# All dimensions should look similar
d <- 8
SG = SGGPcreate(d,201)
SGGPhist(SG)
SGGPhist(SG, ylog=FALSE)

# The first dimension is more active and will have greater depth
SG <- SGGPcreate(d=5, batchsize=10)
SG <- SGGPappend(SGGP=SG, batchsize=100)
SGGPhist(SG)
}
}
