% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGPpredstuff.R
\name{SGGPpred}
\alias{SGGPpred}
\title{Predict}
\usage{
SGGPpred(xp, SG, y, ..., logtheta, theta)
}
\arguments{
\item{xp}{x value to predict at}

\item{SG}{SG object}

\item{y}{Observations for SG}

\item{theta}{Shouldn't this be part of SG?}
}
\value{
Predicted mean values
}
\description{
Predict using SG with y values at xp?
Shouldn't y values already be stored in SG?
}
\examples{
SG <- SGcreate(c(0,0,0), c(1,1,1), batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
SGGPpred(matrix(c(.1,.1,.1),1,3), SG=SG, y=y, c(.1,.1,.1))
cbind(SGGPpred(SG$design, SG=SG, y=y, c(.1,.1,.1))$mean, y) # Should be near equal
}
