% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_pred_fs.R
\name{SGGPpred}
\alias{SGGPpred}
\title{Predict}
\usage{
SGGPpred(xp, SGGP)
}
\arguments{
\item{xp}{x value to predict at}

\item{SGGP}{SG object}
}
\value{
Predicted mean values
}
\description{
Predict using SG with y values at xp?
Shouldn't y values already be stored in SG?
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
SGGPpred(matrix(c(.1,.1,.1),1,3), SG=SG)
cbind(SGGPpred(SGGP$design, SG=SG, y=y, theta=c(.1,.1,.1))$mean, y) # Should be near equal
}
