% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_pred_fs.R
\name{SGGPpred}
\alias{SGGPpred}
\title{Predict}
\usage{
SGGPpred(xp, SGGP)
}
\arguments{
\item{xp}{x value to predict at}

\item{SG}{SG object}

\item{y}{Observations for SG}

\item{theta}{Correlation parameters}

\item{logtheta}{Log of correlation parameters}

\item{...}{Don't use, just forces theta to be named}
}
\value{
Predicted mean values
}
\description{
Predict using SG with y values at xp?
Shouldn't y values already be stored in SG?
}
\examples{
SG <- SGcreate(d=3, batchsize=100)
y <- apply(SGGP$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
SGGPpred(matrix(c(.1,.1,.1),1,3), SG=SG, y=y, theta=c(.1,.1,.1))
cbind(SGGPpred(SGGP$design, SG=SG, y=y, theta=c(.1,.1,.1))$mean, y) # Should be near equal
}
