% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_S3_methods.R, R/SGGP_pred_fs.R
\name{predict.SGGP}
\alias{predict.SGGP}
\alias{SGGPpred}
\title{S3 predict method for SGGP}
\usage{
\method{predict}{SGGP}(object, xp)

SGGPpred(xp, SGGP)
}
\arguments{
\item{object}{SGGP object}

\item{xp}{x value to predict at}

\item{SGGP}{SG object}
}
\value{
Predicted mean values
}
\description{
Passes to SGGPpred

Predict using SG with y values at xp?
Shouldn't y values already be stored in SG?
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
SG <- SGGPfit(SG, Y=y)
SGGPpred(matrix(c(.1,.1,.1),1,3), SGGP=SG)
cbind(SGGPpred(SG$design, SG=SG)$mean, y) # Should be near equal
}
