% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_S3_methods.R, R/SGGP_pred_fs.R
\name{predict.SGGP}
\alias{predict.SGGP}
\alias{SGGPpred}
\title{S3 predict method for SGGP}
\usage{
\method{predict}{SGGP}(object, xp, ...)

SGGPpred(SGGP, xp, fullBayesian = FALSE, theta = NULL,
  outdims = NULL)
}
\arguments{
\item{object}{SGGP object}

\item{xp}{x value to predict at}

\item{...}{Required for S3 consistency}

\item{SGGP}{SG object}

\item{fullBayesian}{Should prediction be done fully Bayesian? Much slower.
Averages over theta samples instead of using thetaMAP.}

\item{theta}{Leave as NULL unless you want to use a value other than thetaMAP.
Much slower.}

\item{outdims}{If multiple outputs fit without PCA and with separate
parameters, you can predict just for certain dimensions to speed it up.
Will leave other columns in the output, but they will be wrong.}
}
\value{
Predicted mean values
}
\description{
Passes to SGGPpred

Predict using SG with y values at xp?
Shouldn't y values already be stored in SG?
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
y <- apply(SG$design, 1, function(x){x[1]+x[2]^2+rnorm(1,0,.01)})
SG <- SGGPfit(SG, Y=y)
SGGPpred(SG, matrix(c(.1,.1,.1),1,3))
cbind(SGGPpred(SG, SG$design)$mean, y) # Should be near equal
}
