% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPvalplot}
\alias{SGGPvalplot}
\title{Plot validation prediction errors for SGGP object}
\usage{
SGGPvalplot(SGGP, Xval, Yval, plot_with = "ggplot2", d = NULL)
}
\arguments{
\item{SGGP}{SGGP object that has been fitted}

\item{Xval}{X validation data}

\item{Yval}{Y validation data}

\item{plot_with}{Should the plot be made with "base" or "ggplot2"?}

\item{d}{If output is multivariate, which column to use. Will do all if
left as NULL.}
}
\value{
None, makes a plot
}
\description{
Plot validation prediction errors for SGGP object
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
f1 <- function(x){x[1]+x[2]^2}
y <- apply(SG$design, 1, f1)
SG <- SGGPfit(SG, y)
Xval <- matrix(runif(3*100), ncol=3)
Yval <- apply(Xval, 1, f1)
SGGPvalplot(SGGP=SG, Xval=Xval, Yval=Yval)
}
