% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPvalstats}
\alias{SGGPvalstats}
\title{Calculate stats for SGGP prediction on validation data}
\usage{
SGGPvalstats(SGGP, Xval, Yval, d = NULL)
}
\arguments{
\item{SGGP}{SGGP object}

\item{Xval}{X validation matrix}

\item{Yval}{Y validation data}

\item{d}{Dimension if multivariate output}
}
\value{
data frame
}
\description{
Calculate stats for SGGP prediction on validation data
}
\examples{
SG <- SGGPcreate(d=3, batchsize=100)
f1 <- function(x){x[1]+x[2]^2}
y <- apply(SG$design, 1, f1)
SG <- SGGPfit(SG, y)
Xval <- matrix(runif(3*100), ncol=3)
Yval <- apply(Xval, 1, f1)
SGGPvalstats(SGGP=SG, Xval=Xval, Yval=Yval)
}
\references{
Gneiting, Tilmann, and Adrian E. Raftery.
"Strictly proper scoring rules, prediction, and estimation."
Journal of the American Statistical Association 102.477 (2007): 359-378.
}
