% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{SGGPvalstats}
\alias{SGGPvalstats}
\title{Calculate stats for SGGP prediction on validation data}
\usage{
SGGPvalstats(SGGP, Xval, Yval, bydim = TRUE, fullBayesian = FALSE, ...)
}
\arguments{
\item{SGGP}{SGGP object}

\item{Xval}{X validation matrix}

\item{Yval}{Y validation data}

\item{bydim}{If multiple outputs, should it be done separately by dimension?}

\item{fullBayesian}{Should prediction be done fully Bayesian? Much slower.
Averages over theta samples instead of using thetaMAP.}

\item{...}{Passed to valstats, such as which stats to calculate.}
}
\value{
data frame
}
\description{
Calculate stats for SGGP prediction on validation data
}
\examples{
\dontrun{
SG <- SGGPcreate(d=3, batchsize=100)
f1 <- function(x){x[1]+x[2]^2}
y <- apply(SG$design, 1, f1)
SG <- SGGPfit(SG, y)
Xval <- matrix(runif(3*100), ncol=3)
Yval <- apply(Xval, 1, f1)
SGGPvalstats(SGGP=SG, Xval=Xval, Yval=Yval)
SGGPvalstats(SGGP=SG, Xval=Xval, Yval=Yval, fullBayesian=TRUE)

# Multiple outputs
SG <- SGGPcreate(d=3, batchsize=100)
f1 <- function(x){x[1]+x[2]^2}
f2 <- function(x){x[1]^1.3+.4*sin(6*x[2])+10}
y1 <- apply(SG$design, 1, f1)#+rnorm(1,0,.01)
y2 <- apply(SG$design, 1, f2)#+rnorm(1,0,.01)
y <- cbind(y1, y2)
SG <- SGGPfit(SG, Y=y)
Xval <- matrix(runif(3*100), ncol=3)
Yval <- cbind(apply(Xval, 1, f1),
              apply(Xval, 1, f2))
SGGPvalstats(SG, Xval, Yval)
SGGPvalstats(SG, Xval, Yval, bydim=FALSE)
}
}
