% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGGP_plot.R
\name{valstats}
\alias{valstats}
\title{Calculate stats for prediction on validation data}
\usage{
valstats(predmean, predvar, Yval)
}
\arguments{
\item{predmean}{Predicted mean}

\item{predvar}{Predicted variance}

\item{Yval}{Y validation data}
}
\value{
data frame
}
\description{
Calculate stats for prediction on validation data
}
\examples{
valstats(c(0,1,2), c(.01,.01,.01), c(0,1.1,1.9))
}
\references{
Gneiting, Tilmann, and Adrian E. Raftery.
"Strictly proper scoring rules, prediction, and estimation."
Journal of the American Statistical Association 102.477 (2007): 359-378.
}
